NanoNotificationBanner: ファイルを分離
